pyproject.toml
[project]
name = "llmops-anime-recommender-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
